refactor: improve code readability and organization in GUI and text model handler
