+++ assignment3/gui.py
@@ -24,20 +24,15 @@ class App(tk.Tk):
         self.title("Group Project Using LLM")
         self.geometry("1000x800")
 
-        # simple state
+        # Application state variables
         self.selected_input = tk.StringVar(value="Text Model")
         self.status_text = tk.StringVar(value="Ready.")
-        self.text_handler = None  # created lazily
-        self.image_handler = None  # created lazily
-        self._img_path = None  # selected image path
-        self._img_preview = None
+        self.text_handler = None  # Lazy initialization
+        self.image_handler = None  # Lazy initialization
+        self._img_path = None  # Selected image path
+        self._img_preview = None  # Image preview for canvas
 
         self._build_ui()
-    
-    # def _on_option_change(self, event):
-    #     self.selected_input = tk.StringVar(value=event.widget.get())
-    #     pass
-        
 
     def _build_ui(self):
         """
@@ -58,7 +53,7 @@ class App(tk.Tk):
         app_container = ttk.Frame(self.tab_app, padding=12)
         app_container.pack(fill=tk.BOTH, expand=True)
 
-        # top controls
+        # Top controls: model selection, browse image, run model
         top_row = ttk.Frame(app_container)
         top_row.pack(fill=tk.X, pady=(0, 8))
 
@@ -71,27 +66,27 @@ class App(tk.Tk):
             textvariable=self.selected_input,
         )
         modelComboBox.pack(side=tk.LEFT, padx=8)
-        # modelComboBox.bind("<<ComboboxSelected>>", self._on_option_change)
+
         value = ""
         if modelComboBox:
             value = modelComboBox.get()
         else:
-            value = ''
-            
+            value = ""
+
         ttk.Button(top_row, text="Browse", command=self._pick_image).pack(
             side=tk.LEFT, padx=(8, 0)
         )
-        ttk.Button(top_row, text="Run Model", command= lambda: self._run_model(value)).pack(
-            side=tk.LEFT, padx=(8, 0)
-        )
+        ttk.Button(
+            top_row, text="Run Model", command=lambda: self._run_model(value)
+        ).pack(side=tk.LEFT, padx=(8, 0))
 
-        # text input
+        # Text input area
         text_group = ttk.LabelFrame(app_container, text="Text Input", padding=8)
         text_group.pack(fill=tk.X, pady=(4, 8))
         self.txt_input = tk.Text(text_group, height=6, wrap="word")
         self.txt_input.pack(fill=tk.X)
 
-        # image preview
+        # Image preview area
         img_group = ttk.LabelFrame(app_container, text="Image Preview", padding=8)
         img_group.pack(fill=tk.X, pady=(4, 8))
         self.img_canvas = tk.Canvas(
@@ -99,13 +94,13 @@ class App(tk.Tk):
         )
         self.img_canvas.pack()
 
-        # output area
+        # Output area
         out_group = ttk.LabelFrame(app_container, text="Output", padding=8)
         out_group.pack(fill=tk.BOTH, expand=True, pady=8)
         self.output = tk.Text(out_group, height=12, wrap="word")
         self.output.pack(fill=tk.BOTH, expand=True)
 
-        # status
+        # Status bar
         status_row = ttk.Frame(app_container)
         status_row.pack(fill=tk.X, pady=(6, 0))
         ttk.Label(status_row, textvariable=self.status_text, anchor="w").pack(
@@ -131,6 +126,7 @@ class App(tk.Tk):
         text = tk.Text(container, wrap="word", height=20)
         text.pack(fill=tk.BOTH, expand=True, pady=8)
 
+        # Display model info from MODEL_INFO
         lines = []
         t = MODEL_INFO.get("text", {})
         i = MODEL_INFO.get("image", {})
@@ -261,9 +257,10 @@ class App(tk.Tk):
             self.status_text.set("Failed. See console for details.")
 
     def _run_model(self, data):
+        """
+        Run the selected model based on user choice.
+        """
         if self.selected_input.get() == "Text Model":
             self._run_text()
-
         if self.selected_input.get() == "Image Model":
             self._run_image()
-        pass
